Separate dependencies from Bodega
PR #20 essentially asks for the ability to use Bodega without using any dependencies. This PR provides a possible implementation for that.

The main idea is that Bodega now exposes two possible libraries to use:

1. BodegaCore - the bulk of the existing functionality, but without the SQLiteStorageEngine
2. Bodega - the existing SQliteStorageEngine implementation

For everyone who wants the simple use-case, they can continue to `import Bodega`. The Bodega target does an `@_exported import BodegaCore`, which means that it re-exposes all the types and symbols of BodegaCore, as if they came from Bodega itself.

For the occasional user who wants to use Bodega *without* dependencies, they would `import BodegaCore` and then implement their own StorageEngine themselves in their own code (such as one that uses the system SQLite3 module as the basis for storage).

The code changes to support this are very minimal:
1. The addition of a new target + product in Package.swift
2. Moving the bulk of the files to a new BodegaCore folder
3. Changing `CacheKey.rawValue` to be public so it can be accessed by the SQLiteStorageEngine
4. Adding a testable import of BodegaCore in a unit test file