Consolidate path concatenation logic
Every usage of `concatenatedPath(key:)` was passing in the `.value` of a `CacheKey`. This introduces the potential for mis-use, if one callsite were to accidentally pass in the `.rawValue` instead.

This can be mitigated by passing the `CacheKey` itself into the method, and then only having a single point of failure. This simplifies usage, because the caller no longer needs to know the specifics of how the URL is constructed.